pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "itrails"
dynamic = ["version"]
description = "Tree reconstruction of ancestry using incomplete lineage sorting"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    
    {name = "David Martin-Pestana", email = "dmape@birc.au.dk"}
]
maintainers = [
    {name = "Iker Rivas-González", email = "iker_rivas_gonzalez@eva.mpg.de"},
    {name = "David Martin-Pestana", email = "dmape@birc.au.dk"}
]
requires-python = ">=3.12"
dependencies = [
  "numpy==1.26.4",
  "scipy==1.13.0",
  "joblib==1.4.2",
  "numba==0.59.1",
  "biopython==1.84",
  "pyyaml==6.0.2",
  "h5py"
]

[project.urls]
Homepage = "https://github.com/trails-phylogeny/itrails"
BugTracker = "https://github.com/trails-phylogeny/itrails/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_file = "src/itrails/_version.py"
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[project.optional-dependencies]
docs = [
    "sphinx",
    "furo",
    "myst-parser",
    "sphinx-autodoc-typehints"
]


[project.scripts]
itrails-optimize = "itrails.workflow_optimize:main"
itrails-viterbi = "itrails.workflow_viterbi:main"
itrails-posterior = "itrails.workflow_posterior:main"
